pyproject.toml
[tool.poetry]
name = "movie"
version = "0.1.0"
description = ""
authors = ["AnOrdinaryUsername <57053268+AnOrdinaryUsername@users.noreply.github.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.2"
djangorestframework = "^3.15.2"
psycopg = "^3.2.3"
djoser = "^2.2.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
